pr comments

+++ resources/shaders/glbillbshader.frag
@@ -24,7 +24,7 @@ float getFogRatio(FogParam fogpar, float dist);
 void main() {
     vec4 fragcol = texture(texture0, texuv);
     uint index = uint(fragcol.r * 255);
-    uvec4 newcol = texelFetch(palbuf, int( 256 * paletteid + index));
+    vec4 newcol = texelFetch(palbuf, int( 256 * paletteid + index));
 
    if (repaint == true) {
 	if (index > 0)
